Move Host to separate file
