+++ lib/music_player/demixer/stem.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:just_audio/just_audio.dart';
-import 'package:musbx/music_player/demixer/demixer_api.dart';
+import 'package:musbx/music_player/demixer/host.dart';
 import 'package:musbx/music_player/music_player.dart';
 
 class Stem {
